PhyloInfoBuilder to new module + docs
Moved PhyloInfoBuilder to a separate file and added docs + doctests.
